Indentated spacing and imported only Scanner & deleted unusual use of this
